Addon-actions: Fix addArgs warnings by refactor
